# Conflicts:
#	ProductPage.cpp

+++ Cakes.cpp
@@ -10,9 +10,9 @@ float Cakes::FinalPrice() const {
     float totalPrice = basePrice * weight;
 
     for (const auto& ornament : ornaments) {
-        totalPrice += static_cast<float>(ornament.getPrice() * ornament.getPieceWeight());
+        totalPrice += ornament.getPrice() * ornament.getPieceWeight();
     }
-    if (days_fresh < 2) {
+    if (static_cast<double>(days_fresh) < 2) {
         totalPrice *= 0.9; // 10% discount
     }
